[ALPS07588156] Coverity: Fix invalid type in argument to printf: 11369436
[Detail]
CID 11369436 (#2 of 2): Invalid type in argument to printf format specifier (PRINTF_ARGS)

[Solution]
Force type conversions to fit format strings

MTK-Commit-Id: ec6a4ceecd207498fc6d677d5bb56237c5a0e148

CR-Id: ALPS07588156
Feature: ARM GPU DDK driver

+++ gpu_mali/mali_valhall/mali-r38p1/drivers/gpu/arm/midgard/device/backend/mali_kbase_device_csf.c
@@ -543,7 +543,7 @@ int kbase_device_firmware_init_once(struct kbase_device *kbdev)
 	snprintf(fw_content, FWLOG_CONTENT_LEN,
 		"[%5lu.%06lu][%llxt]====fwlog End Of File====\n",
 		(unsigned long)ts_nsec,
-		rem_nsec / 1000,
+		(unsigned long)rem_nsec / 1000,
 		mtk_logbuffer_get_timestamp(kbdev, &kbdev->logbuf_regular));
 	memcpy(g_fw_dump_dest, fw_content, FWLOG_CONTENT_LEN);
 #endif /* CONFIG_MALI_MTK_KE_DUMP_FWLOG */
